+++ packages/mongo/mongo_driver.js
@@ -23,6 +23,7 @@ import {
   getAsyncMethodName
 } from "meteor/minimongo/constants";
 import { Meteor } from "meteor/meteor";
+import { ObserveHandle } from './observe_handle';
 
 MongoInternals = {};
